% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateWaferFromDistribution.R
\name{generateWaferFormDistribution}
\alias{generateWaferFormDistribution}
\title{Generate a set of faulty wafer}
\usage{
generateWaferFormDistribution(distributionsList, waferPerDistribution, dieWidth,
  dieHeight)
}
\arguments{
\item{distributionsList:}{a list of matrix. Every element of the matrix is assumed
to represent the probability the chip in that position is faulty.}

\item{waferPerDistribution:}{a vector containing the number of wafer that are to
be generated for every distribution. The length of this vector should match the one
of the distributionsList passed as an argument}

\item{defectAmountRange:}{a vector containing the higher and the
lower number of defects in a wafer}

\item{dieWidth:}{the width of a single chip}

\item{dieHeight:}{the height of a single chip}
}
\value{
a data.frame with 3 colums: the name of the wafer, the X coord of a defect,
the Y coord of a defect. The origin is the lower left corner of the square a wafer
is inscribed in.
}
\description{
Generates a set of wafer filled with good and faulty chips, according to the
fault probability distributions passed as an argument. The number of defect per
wafer is random, but it is ensured it is bind in the range passed as argument.
}

