% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KDEOnWafer.R
\name{KDEOnWafer}
\alias{KDEOnWafer}
\title{Load a data.frame of faults into a matrix}
\usage{
KDEOnWafer(dataFrame, dieWidth, dieHeight, waferRay, plotDistributions = TRUE,
  colorMap = rainbow(20))
}
\arguments{
\item{dayaFrame:}{a data.frame containin all the faults. The format should be the
one specified in the description}

\item{dieWidth:}{the width of the single die}

\item{dieHeigth:}{the heigth of the single die}

\item{waferRay:}{the ray of the wafer}

\item{plotDistributions:}{a boolean flag. If TRUE, all the computed distributions are
plotted.}

\item{colorMap:}{the color map of the plot, if plotDistributions is TRUE.
Must be a valid argument for "col".}
}
\value{
a matrix where any row represent the smoothed fault probability function
on a wafer.
}
\description{
Load data from a data frame. It it assumed that it has three column: the one
named wafer contains the name of the wafer, the one named faultX contains the
x coord of the fault, while the faultY column contains the y coord of the
faulty chip
}

