% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeClassificationRecap.R
\name{writeClassificationRecap}
\alias{writeClassificationRecap}
\title{Perform clustering for wafer data}
\usage{
writeClassificationRecap(faultPositionDataFrame, waferRay, dieWidth, dieHeight,
  splitID = TRUE)
}
\arguments{
\item{faultPositionDataFrame:}{a data frame with three columns: the wafer Id in the first column and the X-Y
coordiantes of the faults in the other two columns}

\item{dieWidth:}{the width of the single die}

\item{dieHeigth:}{the heigth of the single die.}

\item{waferRay:}{the ray of the wafer.}

\item{splitID:}{logical flag. If set to TRUE, the wafer ID is assumed to contain three tokens, separed by "/".
The name is split and in the returned data.frame the single parts are provided.}
}
\value{
A data frame containing the name of the wafer, the number of faulty chips and the cluster.
}
\description{
This function returns a data.frame with the clustering of the data passed as an argument.
Clustering is performed through a SOM network with 3x3 grid.
}

